4.18-4.23

+++ part4/bloglist/controllers/blogs.js
@@ -1,7 +1,8 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
-
+const jwt = require('jsonwebtoken')
+const config = require('../utils/config')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', {name: 1, username: 1})
@@ -9,24 +10,29 @@ blogsRouter.get('/', async (request, response) => {
 })
 
 blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
-
   if(!request.body.title || !request.body.url){
     return response.status(400).end()
   }
 
-  const user = await User.findOne({})
-  blog.user = user._id
+  const user = request.user
 
+  const blog = new Blog(request.body)
+  blog.user = user._id
   const savedBlog = await blog.save()
 
   user.blogs = user.blogs.concat(savedBlog._id)
   await user.save()
-  
   response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) return response.status(404).end()
+
+  if (blog.user.toString() !== request.user.id) {
+    return response.status(403).end()
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
@@ -35,7 +41,6 @@ blogsRouter.put('/:id', async (request, response) => {
   const { title, author, url, likes } = request.body
 
   const blog = await Blog.findById(request.params.id)
-
   if(!blog) return response.status(404).end()
 
   if(!title  || !url) return response.status(400).end()
@@ -46,7 +51,6 @@ blogsRouter.put('/:id', async (request, response) => {
   blog.likes = likes
 
   await blog.save()
-
   response.json(blog)
 })
